Remove password validation
Removing password validation until we add some UI
to specify exactly what is wrong with the password
rather than a generic "not complex enough". This app
is not a high enough target for us to be overly zealous
about password security. If the user wants a simple password
I say let 'em.